Modularized function library. Created two mods: secrets and tables. Split functions based on their use to each of the mods. Also, kept the macros for use in my main lib.rs file. Changed the lines in main that called the original mod to make sure they correctly called mods for each function.
